Moved password verification to authentication module.
